Fix requests made without callback.
Requests made without a callback function would return an empty list.
For example, client.members().request() as mentioned in the docs.

Fixed this by negating the check for including a response in the
return list. And fixed a bug in the ResponseQuery.callback() return
value where it always returns itself (it should return None if there
are no callbacks).